move soc stage to energy stage
